refactor: rename confirmation to confirm
