+++ features/select-doc/internal/src/test/snapshots/images/features.selectdoc.internal.passport.confirm_ConfirmPassportScreenKt_PreviewConfirmPassportScreen_cy.png
